Deepen property-access module: consolidate MOP operations into property.rs
The [[Get]], [[Set]], [[DefineOwnProperty]], [[Delete]], [[HasProperty]],
[[GetOwnProperty]], [[OwnPropertyKeys]], [[GetPrototypeOf]],
[[SetPrototypeOf]], [[IsExtensible]], and [[PreventExtensions]] operations —
one spec-defined module — were smeared across four files and two
abstraction levels:

- types.rs: JsObjectData data-level methods (get_own_property,
  define_own_property, set_property_value — 86 methods)
- mod.rs: thin prototype-chain walkers (get_property_on_id,
  get_property_descriptor_on_id, has_property_on_id)
- eval/modules.rs: get_object_property (the real [[Get]] with proxy
  and module-namespace dispatch) — filed under a file about ES modules
- eval.rs: proxy traps (proxy_set, proxy_define_own_property,
  proxy_delete_property, proxy_get_own_property_descriptor, etc.) and
  array/typed-array exotic operations

This commit consolidates 22 property-access functions into a single
src/interpreter/property.rs module (1,875 lines), providing:

1. Locality: all property-access logic in one navigable file.
   Property bugs now concentrate in one module instead of spreading
   across four files.

2. Canonical entry points: set_object_property, define_object_property,
   and delete_object_property provide unified [[Set]], [[DefineOwnProperty]],
   and [[Delete]] interfaces. Callers can use these instead of manually
   checking for proxies and dispatching between proxy_set and
   set_property_value (currently ~15 inline dispatch sites in eval.rs
   assignment code — migration is a follow-up).

3. Updated AGENTS.md: property.rs added to source layout, stale 'no
   bytecode compilation' note corrected to reflect the existing
   bytecode/ experimental module.

No function signatures, visibility, or logic changed — pure code move.
All 215 cargo tests pass. test262 verification:
- Object built-ins: 6,802 scenarios, 100% pass, 0 regressions
- Proxy: 607 scenarios, 100% pass, 0 regressions
- Expressions: 21,284 scenarios, 100% pass, 0 regressions
