+++ src/interpreter/eval.rs
@@ -15247,1225 +15247,6 @@ impl Interpreter {
         None
     }
 
-    /// Proxy-aware [[Set]] - checks proxy `set` trap, recurses on target if no trap.
-    pub(crate) fn proxy_set(
-        &mut self,
-        obj_id: u64,
-        key: &str,
-        value: JsValue,
-        receiver: &JsValue,
-    ) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(key);
-            match self.invoke_proxy_trap(
-                obj_id,
-                "set",
-                vec![target_val.clone(), key_val, value.clone(), receiver.clone()],
-            ) {
-                Ok(Some(v)) => {
-                    if self.to_boolean_val(&v) {
-                        if let JsValue::Object(ref t) = target_val
-                            && let Some(tobj) = self.get_object(t.id)
-                        {
-                            let target_desc = tobj.borrow().get_own_property(key);
-                            if let Some(ref desc) = target_desc
-                                && desc.configurable == Some(false)
-                            {
-                                if desc.is_data_descriptor()
-                                    && desc.writable == Some(false)
-                                    && !same_value(
-                                        &value,
-                                        desc.value.as_ref().unwrap_or(&JsValue::Undefined),
-                                    )
-                                {
-                                    return Err(self.create_type_error(
-                                            "'set' on proxy: trap returned truish for property which exists in the proxy target as a non-configurable and non-writable data property with a different value",
-                                        ));
-                                }
-                                if desc.is_accessor_descriptor()
-                                    && matches!(
-                                        desc.set.as_ref().unwrap_or(&JsValue::Undefined),
-                                        JsValue::Undefined
-                                    )
-                                {
-                                    return Err(self.create_type_error(
-                                            "'set' on proxy: trap returned truish for property which exists in the proxy target as a non-configurable and non-writable accessor property without a setter",
-                                        ));
-                                }
-                            }
-                        }
-                        Ok(true)
-                    } else {
-                        Ok(false)
-                    }
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_set(t.id, key, value, receiver);
-                    }
-                    Ok(false)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object(obj_id) {
-            // TypedArray [[Set]] §10.4.5.5
-            let is_ta = obj.borrow().typed_array_info().is_some();
-            if is_ta && let Some(index) = canonical_numeric_index_string(key) {
-                let same_val = if let JsValue::Object(ref r) = *receiver {
-                    r.id == obj_id
-                } else {
-                    false
-                };
-                if same_val {
-                    // SameValue(O, Receiver): IntegerIndexedElementSet
-                    let is_bigint = obj
-                        .borrow()
-                        .typed_array_info()
-                        .map(|ta| ta.kind.is_bigint())
-                        .unwrap_or(false);
-                    let num_val = if is_bigint {
-                        self.to_bigint_value(&value)?
-                    } else {
-                        JsValue::Number(self.to_number_value(&value)?)
-                    };
-                    let obj_ref = obj.borrow();
-                    let ta = obj_ref.typed_array_info().unwrap();
-                    if is_valid_integer_index(ta, index) {
-                        let ta_clone = ta.clone();
-                        drop(obj_ref);
-                        typed_array_set_index(&ta_clone, index as usize, &num_val);
-                    }
-                    return Ok(true);
-                } else {
-                    // Different receiver: if invalid index return true without coercing
-                    let valid = {
-                        let obj_ref = obj.borrow();
-                        let ta = obj_ref.typed_array_info().unwrap();
-                        is_valid_integer_index(ta, index)
-                    };
-                    if !valid {
-                        return Ok(true);
-                    }
-                    // Valid index, different receiver: fall through to OrdinarySet below
-                    // OrdinarySet will find writable data descriptor from TypedArray [[GetOwnProperty]],
-                    // then CreateDataProperty(receiver, P, V)
-                }
-            }
-            // OrdinarySetWithOwnDescriptor
-            let own_desc = obj.borrow().get_own_property(key);
-            if let Some(ref desc) = own_desc {
-                if desc.is_accessor_descriptor() {
-                    // Call setter with receiver as this
-                    if let Some(ref setter) = desc.set
-                        && !matches!(setter, JsValue::Undefined)
-                    {
-                        let setter = setter.clone();
-                        match self.call_function(&setter, receiver, &[value]) {
-                            Completion::Normal(_) => return Ok(true),
-                            Completion::Throw(e) => return Err(e),
-                            _ => return Ok(true),
-                        }
-                    }
-                    return Ok(false);
-                }
-                // Data descriptor
-                if desc.writable == Some(false) {
-                    return Ok(false);
-                }
-                // OrdinarySetWithOwnDescriptor step 3.c: use Receiver.[[GetOwnProperty]] / [[DefineOwnProperty]]
-                let recv_id = if let JsValue::Object(r) = receiver {
-                    Some(r.id)
-                } else {
-                    None
-                };
-                if recv_id == Some(obj_id) {
-                    // Common case: receiver is the same object, direct set
-                    return Ok(obj.borrow_mut().set_property_value(key, value));
-                }
-                // Receiver differs: call Receiver.[[GetOwnProperty]](P) and [[DefineOwnProperty]]
-                if let Some(rid) = recv_id {
-                    let existing = self.proxy_get_own_property_descriptor(rid, key)?;
-                    if matches!(existing, JsValue::Undefined) {
-                        // CreateDataProperty(Receiver, P, V)
-                        let desc = crate::interpreter::types::PropertyDescriptor {
-                            value: Some(value),
-                            writable: Some(true),
-                            enumerable: Some(true),
-                            configurable: Some(true),
-                            get: None,
-                            set: None,
-                        };
-                        let desc_val = self.from_property_descriptor(&desc);
-                        return self.proxy_define_own_property(rid, key.to_string(), &desc_val);
-                    } else {
-                        // existingDescriptor found: check accessor or non-writable
-                        let existing_desc = match self.to_property_descriptor(&existing) {
-                            Ok(d) => d,
-                            Err(Some(e)) => return Err(e),
-                            Err(None) => return Ok(false),
-                        };
-                        if existing_desc.is_accessor_descriptor() {
-                            return Ok(false);
-                        }
-                        if existing_desc.writable == Some(false) {
-                            return Ok(false);
-                        }
-                        // [[DefineOwnProperty]](P, {Value: V})
-                        let val_desc = crate::interpreter::types::PropertyDescriptor {
-                            value: Some(value),
-                            writable: None,
-                            enumerable: None,
-                            configurable: None,
-                            get: None,
-                            set: None,
-                        };
-                        let desc_val = self.from_property_descriptor(&val_desc);
-                        return self.proxy_define_own_property(rid, key.to_string(), &desc_val);
-                    }
-                }
-                return Ok(obj.borrow_mut().set_property_value(key, value));
-            }
-            // No own property, walk prototype chain
-            let proto = obj.borrow().prototype_id;
-            if let Some(proto_rc) = proto {
-                let proto_id = proto_rc;
-                return self.proxy_set(proto_id, key, value, receiver);
-            }
-            // No prototype: OrdinarySetWithOwnDescriptor with synthetic {writable:true,...} ownDesc.
-            // Per spec step 1.c.i + 2.c: call Receiver.[[GetOwnProperty]](P) then act on result.
-            if let JsValue::Object(recv_o) = receiver {
-                let recv_id = recv_o.id;
-                let is_proxy_recv = self
-                    .get_object_cell(recv_id)
-                    .is_some_and(|o| o.borrow().is_proxy() || o.borrow().is_proxy_revoked());
-                if is_proxy_recv {
-                    let existing = self.proxy_get_own_property_descriptor(recv_id, key)?;
-                    if matches!(existing, JsValue::Undefined) {
-                        // CreateDataProperty(Receiver, P, V)
-                        let create_desc = crate::interpreter::types::PropertyDescriptor {
-                            value: Some(value),
-                            writable: Some(true),
-                            enumerable: Some(true),
-                            configurable: Some(true),
-                            get: None,
-                            set: None,
-                        };
-                        let desc_val = self.from_property_descriptor(&create_desc);
-                        return self.proxy_define_own_property(recv_id, key.to_string(), &desc_val);
-                    } else {
-                        let existing_desc = match self.to_property_descriptor(&existing) {
-                            Ok(d) => d,
-                            Err(Some(e)) => return Err(e),
-                            Err(None) => return Ok(false),
-                        };
-                        if existing_desc.is_accessor_descriptor() {
-                            return Ok(false);
-                        }
-                        if existing_desc.writable == Some(false) {
-                            return Ok(false);
-                        }
-                        let val_desc = crate::interpreter::types::PropertyDescriptor {
-                            value: Some(value),
-                            writable: None,
-                            enumerable: None,
-                            configurable: None,
-                            get: None,
-                            set: None,
-                        };
-                        let desc_val = self.from_property_descriptor(&val_desc);
-                        return self.proxy_define_own_property(recv_id, key.to_string(), &desc_val);
-                    }
-                }
-                if let Some(recv_obj) = self.get_object_cell(recv_id) {
-                    return Ok(recv_obj.borrow_mut().set_property_value(key, value));
-                }
-            }
-            Ok(obj.borrow_mut().set_property_value(key, value))
-        } else {
-            Ok(false)
-        }
-    }
-
-    pub(crate) fn has_proxy_in_prototype_chain(&self, obj_id: u64) -> bool {
-        let mut current = Some(obj_id);
-        while let Some(id) = current {
-            if self.get_proxy_info(id).is_some() {
-                return true;
-            }
-            let Some(obj) = self.get_object_cell(id) else {
-                return false;
-            };
-            let b = obj.borrow();
-            current = b.prototype_id;
-        }
-        false
-    }
-
-    /// Proxy-aware [[Delete]] - checks proxy `deleteProperty` trap, recurses on target if no trap.
-    pub(crate) fn proxy_delete_property(
-        &mut self,
-        obj_id: u64,
-        key: &str,
-    ) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(key);
-            match self.invoke_proxy_trap(
-                obj_id,
-                "deleteProperty",
-                vec![target_val.clone(), key_val],
-            ) {
-                Ok(Some(v)) => {
-                    let trap_result = self.to_boolean_val(&v);
-                    if trap_result
-                        && let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object_cell(t.id)
-                    {
-                        let target_desc = tobj.borrow().get_own_property(key);
-                        if let Some(ref desc) = target_desc {
-                            if desc.configurable == Some(false) {
-                                return Err(self.create_type_error(
-                                        "'deleteProperty' on proxy: trap returned truish for property which is non-configurable in the proxy target",
-                                    ));
-                            }
-                            if !tobj.borrow().extensible {
-                                return Err(self.create_type_error(
-                                        "'deleteProperty' on proxy: trap returned truish for property but the proxy target is not extensible",
-                                    ));
-                            }
-                        }
-                    }
-                    Ok(trap_result)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_delete_property(t.id, key);
-                    }
-                    Ok(true)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object_cell(obj_id) {
-            // String exotic [[Delete]]: "length" and valid indices are non-configurable
-            {
-                let borrow = obj.borrow();
-                if borrow.class_name == "String"
-                    && let Some(JsValue::String(ref s)) = borrow.primitive_value
-                {
-                    if key == "length" {
-                        return Ok(false);
-                    }
-                    if let Ok(idx) = key.parse::<usize>() {
-                        let char_len = s.to_string().chars().count();
-                        if idx < char_len {
-                            return Ok(false);
-                        }
-                    }
-                }
-            }
-            let mut m = obj.borrow_mut();
-            if let Some(desc) = m.properties.get(key)
-                && desc.configurable == Some(false)
-            {
-                return Ok(false);
-            }
-            m.remove_property(key);
-            if let Some(elems) = m.array_elements_mut()
-                && let Ok(idx) = key.parse::<usize>()
-                && idx < elems.len()
-            {
-                elems[idx] = JsValue::Undefined;
-            }
-            Ok(true)
-        } else {
-            Ok(true)
-        }
-    }
-
-    /// Proxy-aware [[DefineOwnProperty]] - checks proxy `defineProperty` trap, recurses on target if no trap.
-    /// IsCompatiblePropertyDescriptor (§10.1.6.3)
-    fn is_compatible_property_desc(
-        _extensible: bool,
-        desc: &PropertyDescriptor,
-        current: &PropertyDescriptor,
-    ) -> bool {
-        // Step 3: If current.[[Configurable]] is false:
-        if current.configurable == Some(false) {
-            // 3a: If Desc.[[Configurable]] is true, return false
-            if desc.configurable == Some(true) {
-                return false;
-            }
-            // 3b: If Desc has [[Enumerable]] and it differs from current
-            if let Some(desc_enum) = desc.enumerable
-                && current.enumerable != Some(desc_enum)
-            {
-                return false;
-            }
-        }
-        // Step 4: If IsGenericDescriptor(Desc) is true, return true
-        let is_generic = !desc.is_data_descriptor() && !desc.is_accessor_descriptor();
-        if is_generic {
-            return true;
-        }
-        // Step 5: If IsDataDescriptor(current) != IsDataDescriptor(Desc)
-        if current.is_data_descriptor() != desc.is_data_descriptor() {
-            // 5a: If current.[[Configurable]] is false, return false
-            if current.configurable == Some(false) {
-                return false;
-            }
-            return true;
-        }
-        // Step 6: Both are data descriptors
-        if current.is_data_descriptor() && desc.is_data_descriptor() {
-            if current.configurable == Some(false) && current.writable == Some(false) {
-                // 6a.i: If Desc.[[Writable]] is true, return false
-                if desc.writable == Some(true) {
-                    return false;
-                }
-                // 6a.ii: If Desc has [[Value]] and SameValue(Desc.[[Value]], current.[[Value]]) is false
-                if let Some(ref desc_val) = desc.value {
-                    let current_val = current.value.as_ref().unwrap_or(&JsValue::Undefined);
-                    if !same_value(desc_val, current_val) {
-                        return false;
-                    }
-                }
-            }
-            return true;
-        }
-        // Step 7: Both are accessor descriptors
-        if current.configurable == Some(false) {
-            // 7a.i: If Desc has [[Set]] and SameValue(Desc.[[Set]], current.[[Set]]) is false
-            if let Some(ref desc_set) = desc.set {
-                let current_set = current.set.as_ref().unwrap_or(&JsValue::Undefined);
-                if !same_value(desc_set, current_set) {
-                    return false;
-                }
-            }
-            // 7a.ii: If Desc has [[Get]] and SameValue(Desc.[[Get]], current.[[Get]]) is false
-            if let Some(ref desc_get) = desc.get {
-                let current_get = current.get.as_ref().unwrap_or(&JsValue::Undefined);
-                if !same_value(desc_get, current_get) {
-                    return false;
-                }
-            }
-        }
-        true
-    }
-
-    /// §10.4.5.6 TypedArray [[DefineOwnProperty]](P, Desc)
-    /// Returns Ok(None) if not a typed array numeric index (caller should use generic path).
-    /// Returns Ok(Some(bool)) if handled by TypedArray exotic logic.
-    /// Returns Err(JsValue) if an error occurred (e.g., ToBigInt/ToNumber throws).
-    pub(crate) fn typed_array_define_own_property(
-        &mut self,
-        obj_id: u64,
-        key: &str,
-        desc: &PropertyDescriptor,
-    ) -> Result<Option<bool>, JsValue> {
-        use crate::interpreter::types::{canonical_numeric_index_string, is_valid_integer_index};
-
-        let (is_ta, is_bigint) = {
-            if let Some(obj) = self.get_object_cell(obj_id) {
-                let b = obj.borrow();
-                if let Some(ta) = b.typed_array_info() {
-                    (true, ta.kind.is_bigint())
-                } else {
-                    (false, false)
-                }
-            } else {
-                (false, false)
-            }
-        };
-
-        if !is_ta {
-            return Ok(None);
-        }
-
-        let index = match canonical_numeric_index_string(key) {
-            Some(idx) => idx,
-            None => return Ok(None),
-        };
-
-        // §10.4.5.6 step 3b.i: if not a valid integer index, return false
-        {
-            let valid = if let Some(obj) = self.get_object_cell(obj_id) {
-                let b = obj.borrow();
-                b.typed_array_info()
-                    .as_ref()
-                    .map(|ta| is_valid_integer_index(ta, index))
-                    .unwrap_or(false)
-            } else {
-                false
-            };
-            if !valid {
-                return Ok(Some(false));
-            }
-        }
-
-        // §10.4.5.6 step 3b.ii: accessor descriptors not allowed
-        if desc.get.is_some() || desc.set.is_some() {
-            return Ok(Some(false));
-        }
-        // §10.4.5.6 step 3b.iii: configurable must not be false
-        if desc.configurable == Some(false) {
-            return Ok(Some(false));
-        }
-        // §10.4.5.6 step 3b.iv: enumerable must not be false
-        if desc.enumerable == Some(false) {
-            return Ok(Some(false));
-        }
-        // §10.4.5.6 step 3b.v: writable must not be false
-        if desc.writable == Some(false) {
-            return Ok(Some(false));
-        }
-
-        // §10.4.5.6 step 3b.vi: if [[Value]] present, call IntegerIndexedElementSet
-        if let Some(ref value) = desc.value {
-            // IntegerIndexedElementSet: ToNumber/ToBigInt first (may throw), then check valid
-            let num_val = if is_bigint {
-                self.to_bigint_value(value)?
-            } else {
-                JsValue::Number(self.to_number_value(value)?)
-            };
-            // After conversion, re-read ta info (buffer may have been detached during conversion)
-            if let Some(obj) = self.get_object_cell(obj_id) {
-                let b = obj.borrow();
-                if let Some(ta) = b.typed_array_info()
-                    && is_valid_integer_index(ta, index)
-                {
-                    let ta_clone2 = ta.clone();
-                    drop(b);
-                    crate::interpreter::types::typed_array_set_index(
-                        &ta_clone2,
-                        index as usize,
-                        &num_val,
-                    );
-                }
-            }
-            return Ok(Some(true));
-        }
-
-        Ok(Some(true))
-    }
-
-    pub(crate) fn proxy_define_own_property(
-        &mut self,
-        obj_id: u64,
-        key: String,
-        desc_val: &JsValue,
-    ) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(&key);
-            match self.invoke_proxy_trap(
-                obj_id,
-                "defineProperty",
-                vec![target_val.clone(), key_val, desc_val.clone()],
-            ) {
-                Ok(Some(v)) => {
-                    let trap_result = self.to_boolean_val(&v);
-                    if !trap_result {
-                        return Ok(false);
-                    }
-                    if let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object_cell(t.id)
-                    {
-                        let target_desc = tobj.borrow().get_own_property(&key);
-                        let target_extensible = tobj.borrow().extensible;
-                        let desc = self.to_property_descriptor(desc_val).ok();
-                        let setting_config_false =
-                            desc.as_ref().is_some_and(|d| d.configurable == Some(false));
-
-                        if let Some(ref desc) = desc {
-                            // Step 19: targetDesc is undefined
-                            if target_desc.is_none() {
-                                if !target_extensible {
-                                    return Err(self.create_type_error(
-                                        "'defineProperty' on proxy: trap returned truish for adding property to the non-extensible proxy target",
-                                    ));
-                                }
-                                if setting_config_false {
-                                    return Err(self.create_type_error(
-                                        "'defineProperty' on proxy: trap returned truish for defining non-configurable property which does not exist on the proxy target",
-                                    ));
-                                }
-                            }
-                            // Step 20: targetDesc is not undefined
-                            if let Some(ref td) = target_desc {
-                                // 20a: IsCompatiblePropertyDescriptor check
-                                if !Self::is_compatible_property_desc(target_extensible, desc, td) {
-                                    return Err(self.create_type_error(
-                                        "'defineProperty' on proxy: trap returned truish for property descriptor not compatible with the existing property in the proxy target",
-                                    ));
-                                }
-                                // 20b: settingConfigFalse + target configurable
-                                if setting_config_false && td.configurable == Some(true) {
-                                    return Err(self.create_type_error(
-                                        "'defineProperty' on proxy: trap returned truish for defining non-configurable property which is configurable in the proxy target",
-                                    ));
-                                }
-                                // 20c: target non-configurable+writable, desc says non-writable
-                                if td.is_data_descriptor()
-                                    && td.configurable == Some(false)
-                                    && td.writable == Some(true)
-                                    && desc.writable == Some(false)
-                                {
-                                    return Err(self.create_type_error(
-                                        "'defineProperty' on proxy: trap returned truish for setting non-writable on a non-configurable writable property in the proxy target",
-                                    ));
-                                }
-                            }
-                        }
-                    }
-                    Ok(true)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_define_own_property(t.id, key, desc_val);
-                    }
-                    Ok(false)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object_cell(obj_id) {
-            // Deferred namespace: trigger evaluation on [[DefineOwnProperty]] with non-symbol-like key
-            {
-                let is_deferred_ns = obj
-                    .borrow()
-                    .module_namespace()
-                    .as_ref()
-                    .is_some_and(|ns| ns.deferred);
-                if is_deferred_ns && !Self::is_symbol_like_namespace_key(&key, true) {
-                    self.ensure_deferred_namespace_evaluation(obj_id)?;
-                }
-            }
-            let obj = self.get_object(obj_id).unwrap();
-            let is_array = obj.borrow().class_name == "Array";
-            match self.to_property_descriptor(desc_val) {
-                Ok(desc) => {
-                    if is_array {
-                        self.array_define_own_property(obj_id as usize, &key, desc)
-                    } else {
-                        Ok(obj.borrow_mut().define_own_property(key, desc))
-                    }
-                }
-                Err(Some(e)) => Err(e),
-                Err(None) => Ok(false),
-            }
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// Proxy-aware [[GetOwnProperty]] - checks proxy `getOwnPropertyDescriptor` trap, recurses on target if no trap.
-    pub(crate) fn proxy_get_own_property_descriptor(
-        &mut self,
-        obj_id: u64,
-        key: &str,
-    ) -> Result<JsValue, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(key);
-            match self.invoke_proxy_trap(
-                obj_id,
-                "getOwnPropertyDescriptor",
-                vec![target_val.clone(), key_val],
-            ) {
-                Ok(Some(v)) => {
-                    // Step 11: If Type(trapResultObj) is neither Object nor Undefined, throw TypeError
-                    if !matches!(v, JsValue::Object(_) | JsValue::Undefined) {
-                        return Err(self.create_type_error(
-                            "'getOwnPropertyDescriptor' on proxy: trap returned neither Object nor undefined",
-                        ));
-                    }
-                    if let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object_cell(t.id)
-                    {
-                        let target_desc = tobj.borrow().get_own_property(key);
-                        let target_extensible = tobj.borrow().extensible;
-                        if matches!(v, JsValue::Undefined) {
-                            if let Some(ref td) = target_desc {
-                                if td.configurable == Some(false) {
-                                    return Err(self.create_type_error(
-                                        "'getOwnPropertyDescriptor' on proxy: trap returned undefined for property which is non-configurable in the proxy target",
-                                    ));
-                                }
-                                if !target_extensible {
-                                    return Err(self.create_type_error(
-                                        "'getOwnPropertyDescriptor' on proxy: trap returned undefined for property which exists in the non-extensible proxy target",
-                                    ));
-                                }
-                            }
-                        } else if matches!(v, JsValue::Object(_)) {
-                            let result_desc = match self.to_property_descriptor(&v) {
-                                Ok(d) => d,
-                                Err(Some(e)) => return Err(e),
-                                Err(None) => return Ok(JsValue::Undefined),
-                            };
-                            // Step 22: If resultDesc.[[Configurable]] is false
-                            if result_desc.configurable == Some(false) {
-                                // 22a: If targetDesc is undefined or targetDesc.[[Configurable]] is true
-                                if target_desc.is_none()
-                                    || target_desc
-                                        .as_ref()
-                                        .is_some_and(|td| td.configurable == Some(true))
-                                {
-                                    return Err(self.create_type_error(
-                                            "'getOwnPropertyDescriptor' on proxy: trap reported non-configurable for a property that is either non-existent or configurable in the proxy target",
-                                        ));
-                                }
-                            }
-
-                            if let Some(ref td) = target_desc {
-                                if td.configurable == Some(false) {
-                                    // Step 21a: resultDesc configurable:true for non-configurable target
-                                    if result_desc.configurable == Some(true) {
-                                        return Err(self.create_type_error(
-                                                "'getOwnPropertyDescriptor' on proxy: trap returned descriptor with configurable: true for non-configurable property in the proxy target",
-                                            ));
-                                    }
-                                    // Step 21b: writable:true for non-configurable non-writable target
-                                    if td.is_data_descriptor()
-                                        && td.writable == Some(false)
-                                        && result_desc.writable == Some(true)
-                                    {
-                                        return Err(self.create_type_error(
-                                                "'getOwnPropertyDescriptor' on proxy: trap returned descriptor with writable: true for non-configurable non-writable property in the proxy target",
-                                            ));
-                                    }
-                                    // Enumerable must match for non-configurable target
-                                    if result_desc.enumerable.is_some()
-                                        && result_desc.enumerable != td.enumerable
-                                    {
-                                        return Err(self.create_type_error(
-                                                "'getOwnPropertyDescriptor' on proxy: trap returned descriptor with incompatible enumerable for non-configurable property in the proxy target",
-                                            ));
-                                    }
-                                    // Type mismatch: data vs accessor
-                                    if td.is_data_descriptor() != result_desc.is_data_descriptor()
-                                        && td.is_accessor_descriptor()
-                                            != result_desc.is_accessor_descriptor()
-                                    {
-                                        return Err(self.create_type_error(
-                                                "'getOwnPropertyDescriptor' on proxy: trap returned descriptor with different type than non-configurable property in the proxy target",
-                                            ));
-                                    }
-                                    // Non-configurable accessor: getter/setter must match
-                                    if td.is_accessor_descriptor()
-                                        && result_desc.is_accessor_descriptor()
-                                    {
-                                        let td_get = td.get.as_ref();
-                                        let rd_get = result_desc.get.as_ref();
-                                        if rd_get.is_some()
-                                            && !Self::same_value_option(td_get, rd_get)
-                                        {
-                                            return Err(self.create_type_error(
-                                                    "'getOwnPropertyDescriptor' on proxy: trap returned descriptor with different getter for non-configurable property in the proxy target",
-                                                ));
-                                        }
-                                        let td_set = td.set.as_ref();
-                                        let rd_set = result_desc.set.as_ref();
-                                        if rd_set.is_some()
-                                            && !Self::same_value_option(td_set, rd_set)
-                                        {
-                                            return Err(self.create_type_error(
-                                                    "'getOwnPropertyDescriptor' on proxy: trap returned descriptor with different setter for non-configurable property in the proxy target",
-                                                ));
-                                        }
-                                    }
-                                    // Non-configurable non-writable data: value must match
-                                    if td.is_data_descriptor()
-                                        && td.writable == Some(false)
-                                        && result_desc.is_data_descriptor()
-                                        && result_desc.writable != Some(true)
-                                        && let (Some(tv), Some(rv)) =
-                                            (&td.value, &result_desc.value)
-                                        && !crate::interpreter::helpers::same_value(tv, rv)
-                                    {
-                                        return Err(self.create_type_error(
-                                            "'getOwnPropertyDescriptor' on proxy: trap returned descriptor with different value for non-configurable non-writable property in the proxy target",
-                                        ));
-                                    }
-                                    // Step 21b: non-configurable non-writable result but writable target
-                                    if result_desc.is_data_descriptor()
-                                        && result_desc.writable == Some(false)
-                                        && td.writable == Some(true)
-                                    {
-                                        return Err(self.create_type_error(
-                                                "'getOwnPropertyDescriptor' on proxy: trap returned non-configurable non-writable descriptor for a configurable or writable property in the proxy target",
-                                            ));
-                                    }
-                                }
-                            } else if !target_extensible {
-                                return Err(self.create_type_error(
-                                        "'getOwnPropertyDescriptor' on proxy: trap returned descriptor for property which does not exist in the non-extensible proxy target",
-                                    ));
-                            }
-                            return Ok(self.from_property_descriptor(&result_desc));
-                        }
-                    }
-                    Ok(v)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_get_own_property_descriptor(t.id, key);
-                    }
-                    Ok(JsValue::Undefined)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object(obj_id) {
-            // §10.4.6.4 [[GetOwnProperty]] step 4: namespace [[Get]] can throw for TDZ
-            self.check_namespace_tdz(obj_id, key)?;
-            let desc = obj.borrow().get_own_property(key);
-            match desc {
-                Some(d) => Ok(self.from_property_descriptor(&d)),
-                None => Ok(JsValue::Undefined),
-            }
-        } else {
-            Ok(JsValue::Undefined)
-        }
-    }
-
-    /// Proxy-aware [[OwnPropertyKeys]] - checks proxy `ownKeys` trap, recurses on target if no trap.
-    /// Returns all own property keys (for getOwnPropertyNames).
-    pub(crate) fn proxy_own_keys(&mut self, obj_id: u64) -> Result<Vec<JsValue>, JsValue> {
-        const MAX_PROXY_OWNKEYS_RESULT_LEN: usize = 1_000_000;
-
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            match self.invoke_proxy_trap(obj_id, "ownKeys", vec![target_val.clone()]) {
-                Ok(Some(v)) => {
-                    if !matches!(v, JsValue::Object(_)) {
-                        return Err(
-                            self.create_type_error("CreateListFromArrayLike called on non-object")
-                        );
-                    }
-                    if let JsValue::Object(arr) = &v {
-                        let arr_id = arr.id;
-                        // Use [[Get]] for length (spec: CreateListFromArrayLike)
-                        let len_val = match self.get_object_property(arr_id, "length", &v) {
-                            Completion::Normal(lv) => lv,
-                            Completion::Throw(e) => return Err(e),
-                            _ => JsValue::Undefined,
-                        };
-                        let len = match len_val {
-                            JsValue::Number(n) if n.is_finite() && n > 0.0 => {
-                                let len = n.floor() as usize;
-                                if len > MAX_PROXY_OWNKEYS_RESULT_LEN {
-                                    return Err(self.create_type_error(
-                                        "'ownKeys' on proxy: trap result length exceeds supported limit",
-                                    ));
-                                }
-                                len
-                            }
-                            JsValue::Number(_) => 0,
-                            _ => {
-                                return Err(self.create_type_error(
-                                    "ownKeys trap result length is not a number",
-                                ));
-                            }
-                        };
-                        // Use [[Get]] for each element
-                        let mut keys: Vec<JsValue> = Vec::with_capacity(len);
-                        for i in 0..len {
-                            let elem = match self.get_object_property(arr_id, &i.to_string(), &v) {
-                                Completion::Normal(ev) => ev,
-                                Completion::Throw(e) => return Err(e),
-                                _ => JsValue::Undefined,
-                            };
-                            keys.push(elem);
-                        }
-                        for key in &keys {
-                            if !matches!(key, JsValue::String(_) | JsValue::Symbol(_)) {
-                                return Err(self.create_type_error(
-                                    "'ownKeys' on proxy: trap returned non-string/symbol key",
-                                ));
-                            }
-                        }
-                        let mut seen = HashSet::new();
-                        for key in &keys {
-                            let key_str = to_property_key_string(key);
-                            if !seen.insert(key_str) {
-                                return Err(self.create_type_error(
-                                    "'ownKeys' on proxy: trap returned duplicate entries",
-                                ));
-                            }
-                        }
-                        self.validate_ownkeys_invariant(&v, &target_val)?;
-                        Ok(keys)
-                    } else {
-                        Ok(vec![])
-                    }
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_own_keys(t.id);
-                    }
-                    Ok(vec![])
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object_cell(obj_id) {
-            // Deferred namespace: trigger evaluation on [[OwnPropertyKeys]]
-            {
-                let is_deferred_ns = obj
-                    .borrow()
-                    .module_namespace()
-                    .as_ref()
-                    .is_some_and(|ns| ns.deferred);
-                if is_deferred_ns {
-                    self.ensure_deferred_namespace_evaluation(obj_id)?;
-                }
-            }
-            // OrdinaryOwnPropertyKeys: integer indices (sorted), then string keys (in creation order), then symbol keys
-            let obj = self.get_object(obj_id).unwrap();
-            let b = obj.borrow();
-
-            // String exotic objects (§10.4.3.3): virtual char indices included
-            let is_string_wrapper =
-                b.class_name == "String" && matches!(b.primitive_value, Some(JsValue::String(_)));
-            let string_len = if is_string_wrapper {
-                if let Some(JsValue::String(ref s)) = b.primitive_value {
-                    s.code_units.len()
-                } else {
-                    0
-                }
-            } else {
-                0
-            };
-
-            let mut int_keys_set: std::collections::BTreeMap<u64, String> =
-                std::collections::BTreeMap::new();
-            let mut str_keys: Vec<String> = Vec::new();
-            let mut sym_keys: Vec<String> = Vec::new();
-
-            // String exotic: char indices 0..len are virtual integer indices
-            if is_string_wrapper {
-                for i in 0..string_len {
-                    int_keys_set.insert(i as u64, i.to_string());
-                }
-            }
-
-            // TypedArray [[OwnPropertyKeys]]: virtual integer indices
-            if let Some(ta) = b.typed_array_info() {
-                use crate::interpreter::types::{is_typed_array_out_of_bounds, typed_array_length};
-                if !is_typed_array_out_of_bounds(ta) {
-                    let len = typed_array_length(ta);
-                    for i in 0..len {
-                        int_keys_set.insert(i as u64, i.to_string());
-                    }
-                }
-            }
-
-            if let Some(elems) = b.array_elements() {
-                for (i, value) in elems.iter().enumerate() {
-                    if matches!(value, JsValue::Undefined) || i > 0xFFFF_FFFE {
-                        continue;
-                    }
-                    int_keys_set.insert(i as u64, i.to_string());
-                }
-            }
-
-            for k in &b.property_order {
-                if k.starts_with("Symbol(") {
-                    sym_keys.push(k.to_string());
-                } else if let Ok(n) = k.parse::<u64>() {
-                    if n.to_string() == **k {
-                        // This is an integer index - add/overwrite (string char indices take precedence, but we let btreemap handle uniqueness)
-                        int_keys_set.insert(n, k.to_string());
-                    } else {
-                        str_keys.push(k.to_string());
-                    }
-                } else {
-                    // Skip "length" for string wrappers - it's virtual, added separately
-                    if is_string_wrapper && &**k == "length" {
-                        continue;
-                    }
-                    str_keys.push(k.to_string());
-                }
-            }
-
-            let mut result: Vec<JsValue> = Vec::new();
-            for (_, k) in int_keys_set {
-                result.push(JsValue::String(JsString::from_str(&k)));
-            }
-            for k in str_keys {
-                result.push(JsValue::String(JsString::from_str(&k)));
-            }
-            // String exotic: "length" is a virtual non-enumerable string key (after other str keys, before symbols)
-            if is_string_wrapper {
-                result.push(JsValue::String(JsString::from_str("length")));
-            }
-            for k in sym_keys {
-                result.push(self.symbol_key_to_jsvalue(&k));
-            }
-            Ok(result)
-        } else {
-            Ok(vec![])
-        }
-    }
-
-    /// Proxy-aware enumerable keys with prototype chain walk for for-in loops.
-    pub(crate) fn proxy_enumerable_keys_with_proto(
-        &mut self,
-        obj_id: u64,
-    ) -> Result<Vec<String>, JsValue> {
-        let mut seen = HashSet::new();
-        let mut keys = Vec::new();
-        let mut current_id = Some(obj_id);
-
-        while let Some(cid) = current_id {
-            // Get own keys for current object (proxy-aware)
-            let own_keys = self.proxy_own_keys(cid)?;
-            for key in &own_keys {
-                if let JsValue::String(s) = key {
-                    let key_str = s.to_rust_string();
-                    if key_str.starts_with("Symbol(") {
-                        continue;
-                    }
-                    if seen.contains(&key_str) {
-                        continue;
-                    }
-                    // Check enumerability via proxy-aware [[GetOwnProperty]]
-                    {
-                        let desc_val = self.proxy_get_own_property_descriptor(cid, &key_str)?;
-                        seen.insert(key_str.clone());
-                        if !matches!(desc_val, JsValue::Undefined)
-                            && let Ok(desc) = self.to_property_descriptor(&desc_val)
-                            && desc.enumerable != Some(false)
-                        {
-                            keys.push(key_str);
-                        }
-                    }
-                }
-            }
-
-            // Walk prototype chain (proxy-aware)
-            match self.proxy_get_prototype_of(cid) {
-                Ok(JsValue::Object(proto_ref)) => {
-                    current_id = Some(proto_ref.id);
-                }
-                Ok(_) => current_id = None,
-                Err(e) => return Err(e),
-            }
-        }
-        Ok(keys)
-    }
-
-    /// Proxy-aware [[GetPrototypeOf]] - checks proxy `getPrototypeOf` trap, recurses on target if no trap.
-    pub(crate) fn proxy_get_prototype_of(&mut self, obj_id: u64) -> Result<JsValue, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            match self.invoke_proxy_trap(obj_id, "getPrototypeOf", vec![target_val.clone()]) {
-                Ok(Some(v)) => {
-                    if !matches!(v, JsValue::Object(_) | JsValue::Null) {
-                        return Err(self.create_type_error(
-                            "'getPrototypeOf' on proxy: trap returned neither object nor null",
-                        ));
-                    }
-                    // Step 8: extensibleTarget = ? IsExtensible(target)
-                    if let JsValue::Object(ref t) = target_val {
-                        let extensible_target = self.proxy_is_extensible(t.id)?;
-                        if !extensible_target {
-                            // Step 10: targetProto = ? target.[[GetPrototypeOf]]()
-                            let actual_proto = self.proxy_get_prototype_of(t.id)?;
-                            let same = match (&v, &actual_proto) {
-                                (JsValue::Object(a), JsValue::Object(b)) => a.id == b.id,
-                                (JsValue::Null, JsValue::Null) => true,
-                                _ => false,
-                            };
-                            if !same {
-                                return Err(self.create_type_error(
-                                    "'getPrototypeOf' on proxy: proxy target is non-extensible but the trap did not return its actual prototype",
-                                ));
-                            }
-                        }
-                    }
-                    Ok(v)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_get_prototype_of(t.id);
-                    }
-                    Ok(JsValue::Null)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object_cell(obj_id) {
-            if let Some(id) = obj.borrow().prototype_id {
-                Ok(JsValue::Object(crate::types::JsObject { id }))
-            } else {
-                Ok(JsValue::Null)
-            }
-        } else {
-            Ok(JsValue::Null)
-        }
-    }
-
-    /// Proxy-aware [[SetPrototypeOf]] - checks proxy `setPrototypeOf` trap, recurses on target if no trap.
-    pub(crate) fn proxy_set_prototype_of(
-        &mut self,
-        obj_id: u64,
-        proto: &JsValue,
-    ) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            match self.invoke_proxy_trap(
-                obj_id,
-                "setPrototypeOf",
-                vec![target_val.clone(), proto.clone()],
-            ) {
-                Ok(Some(v)) => {
-                    if !self.to_boolean_val(&v) {
-                        return Ok(false);
-                    }
-                    // Step 8: IsExtensible(target) — may throw, must use proxy-aware check
-                    let target_id = if let JsValue::Object(ref t) = target_val {
-                        t.id
-                    } else {
-                        return Ok(true);
-                    };
-                    let extensible_target = self.proxy_is_extensible(target_id)?;
-                    // Step 9: if extensible, no invariant to check
-                    if extensible_target {
-                        return Ok(true);
-                    }
-                    // Step 10: GetPrototypeOf(target) — may throw
-                    let actual_proto = self.proxy_get_prototype_of(target_id)?;
-                    let same = match (proto, &actual_proto) {
-                        (JsValue::Object(a), JsValue::Object(b)) => a.id == b.id,
-                        (JsValue::Null, JsValue::Null) => true,
-                        _ => false,
-                    };
-                    if !same {
-                        return Err(self.create_type_error(
-                            "'setPrototypeOf' on proxy: trap returned truish for setting a new prototype on the non-extensible proxy target",
-                        ));
-                    }
-                    Ok(true)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_set_prototype_of(t.id, proto);
-                    }
-                    Ok(true)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object_cell(obj_id) {
-            // OrdinarySetPrototypeOf
-            let current_proto_id = obj.borrow().prototype_id;
-            let new_proto_id = if let JsValue::Object(p) = proto {
-                Some(p.id)
-            } else {
-                None
-            };
-            let same = (matches!(proto, JsValue::Null) && current_proto_id.is_none())
-                || matches!((new_proto_id, current_proto_id), (Some(a), Some(b)) if a == b);
-            if same {
-                return Ok(true);
-            }
-            if !obj.borrow().extensible {
-                return Ok(false);
-            }
-            // Cycle check
-            if let JsValue::Object(p) = proto {
-                let mut check_id = Some(p.id);
-                while let Some(cid) = check_id {
-                    if cid == obj_id {
-                        return Ok(false);
-                    }
-                    check_id = self
-                        .get_object_cell(cid)
-                        .and_then(|o| o.borrow().prototype_id.as_ref().copied());
-                }
-            }
-            match proto {
-                JsValue::Null => {
-                    obj.borrow_mut().prototype_id = None;
-                }
-                JsValue::Object(p) => {
-                    if let Some(po) = self.get_object_cell(p.id) {
-                        obj.borrow_mut().prototype_id = Some(po.borrow().id.unwrap());
-                    }
-                }
-                _ => {}
-            }
-            Ok(true)
-        } else {
-            Ok(true)
-        }
-    }
-
-    /// Proxy-aware [[IsExtensible]] - checks proxy `isExtensible` trap, recurses on target if no trap.
-    pub(crate) fn proxy_is_extensible(&mut self, obj_id: u64) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            match self.invoke_proxy_trap(obj_id, "isExtensible", vec![target_val.clone()]) {
-                Ok(Some(v)) => {
-                    let trap_result = self.to_boolean_val(&v);
-                    if let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object_cell(t.id)
-                    {
-                        let target_extensible = tobj.borrow().extensible;
-                        if trap_result != target_extensible {
-                            return Err(self.create_type_error(
-                                "'isExtensible' on proxy: trap result does not reflect extensibility of proxy target",
-                            ));
-                        }
-                    }
-                    Ok(trap_result)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_is_extensible(t.id);
-                    }
-                    Ok(false)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object_cell(obj_id) {
-            Ok(obj.borrow().extensible)
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// Proxy-aware [[PreventExtensions]] - checks proxy `preventExtensions` trap, recurses on target if no trap.
-    pub(crate) fn proxy_prevent_extensions(&mut self, obj_id: u64) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            match self.invoke_proxy_trap(obj_id, "preventExtensions", vec![target_val.clone()]) {
-                Ok(Some(v)) => {
-                    let trap_result = self.to_boolean_val(&v);
-                    if trap_result
-                        && let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object_cell(t.id)
-                        && tobj.borrow().extensible
-                    {
-                        return Err(self.create_type_error(
-                                "'preventExtensions' on proxy: trap returned truish but the proxy target is extensible",
-                            ));
-                    }
-                    Ok(trap_result)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_prevent_extensions(t.id);
-                    }
-                    Ok(false)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object_cell(obj_id) {
-            obj.borrow_mut().extensible = false;
-            Ok(true)
-        } else {
-            Ok(false)
-        }
-    }
 
     /// Get the super base object ID from __home_object__.__proto__ in the given env.
     /// Returns Ok(Some(id)) for a valid super base, Ok(None) for null prototype, or
