+++ src/interpreter/eval/modules.rs
@@ -3,7 +3,7 @@ use super::*;
 impl Interpreter {
     /// Resolve a module export value by following re-export chains recursively.
     /// Used by namespace [[Get]] to dynamically resolve live bindings.
-    pub(super) fn resolve_module_export_value(
+    pub(crate) fn resolve_module_export_value(
         &mut self,
         binding_name: &str,
         env: &crate::interpreter::types::EnvRef,
@@ -249,268 +249,6 @@ impl Interpreter {
                 }
                 Err(e.clone())
             }
-        }
-    }
-
-    pub(crate) fn get_object_property(
-        &mut self,
-        obj_id: u64,
-        key: &str,
-        this_val: &JsValue,
-    ) -> Completion {
-        // Single-borrow fast path: classify object and check own property in one borrow
-        if let Some(obj) = self.get_object_cell(obj_id) {
-            let b = obj.borrow();
-            let is_proxy = b.proxy().is_some();
-            let has_module_ns = b.module_namespace().is_some();
-
-            if !is_proxy && !has_module_ns {
-                // Fast path for ordinary objects (the common case)
-                let is_ta = b.typed_array_info().is_some();
-
-                // TypedArray: canonical numeric index strings must not walk prototype
-                if is_ta
-                    && let Some(index) =
-                        crate::interpreter::types::canonical_numeric_index_string(key)
-                {
-                    use crate::interpreter::types::{
-                        is_valid_integer_index, typed_array_get_index,
-                    };
-                    let ta = b.typed_array_info().unwrap();
-                    if is_valid_integer_index(ta, index) {
-                        return Completion::Normal(typed_array_get_index(ta, index as usize));
-                    }
-                    return Completion::Normal(JsValue::Undefined);
-                }
-
-                // Check own property
-                let own_desc = b.get_own_property_full(key);
-                match own_desc {
-                    Some(ref d)
-                        if d.get.is_some() && !matches!(d.get, Some(JsValue::Undefined)) =>
-                    {
-                        let getter = d.get.clone().unwrap();
-                        drop(b);
-                        return self.call_function(&getter, this_val, &[]);
-                    }
-                    Some(ref d) if d.get.is_some() => {
-                        return Completion::Normal(JsValue::Undefined);
-                    }
-                    Some(ref d) => {
-                        return Completion::Normal(d.value.clone().unwrap_or(JsValue::Undefined));
-                    }
-                    None => {}
-                }
-
-                // Own property not found — walk prototype chain
-                let proto = b.prototype_id;
-                drop(b);
-                if let Some(proto_rc) = proto {
-                    let proto_id = proto_rc;
-                    return self.get_object_property(proto_id, key, this_val);
-                }
-                return Completion::Normal(JsValue::Undefined);
-            }
-            drop(b);
-        }
-
-        // Slow path: proxy or module namespace objects
-        self.get_object_property_slow(obj_id, key, this_val)
-    }
-
-    fn get_object_property_slow(
-        &mut self,
-        obj_id: u64,
-        key: &str,
-        this_val: &JsValue,
-    ) -> Completion {
-        // Check if object is a proxy
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(key);
-            let receiver = this_val.clone();
-            match self.invoke_proxy_trap(obj_id, "get", vec![target_val.clone(), key_val, receiver])
-            {
-                Ok(Some(v)) => {
-                    // Invariant checks
-                    if let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object_cell(t.id)
-                    {
-                        let target_desc = tobj.borrow().get_own_property(key);
-                        if let Some(ref desc) = target_desc
-                            && desc.configurable == Some(false)
-                        {
-                            if desc.is_data_descriptor()
-                                && desc.writable == Some(false)
-                                && !same_value(
-                                    &v,
-                                    desc.value.as_ref().unwrap_or(&JsValue::Undefined),
-                                )
-                            {
-                                return Completion::Throw(self.create_type_error(
-                                        "'get' on proxy: property is a read-only and non-configurable data property on the proxy target but the proxy did not return its actual value",
-                                    ));
-                            }
-                            if desc.is_accessor_descriptor()
-                                && matches!(
-                                    desc.get.as_ref().unwrap_or(&JsValue::Undefined),
-                                    JsValue::Undefined
-                                )
-                                && !matches!(v, JsValue::Undefined)
-                            {
-                                return Completion::Throw(self.create_type_error(
-                                        "'get' on proxy: property is a non-configurable accessor property on the proxy target and does not have a getter function, but the trap did not return 'undefined'",
-                                    ));
-                            }
-                        }
-                    }
-                    return Completion::Normal(v);
-                }
-                Ok(None) => {
-                    // No trap, fall through to target
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.get_object_property(t.id, key, this_val);
-                    }
-                    return Completion::Normal(JsValue::Undefined);
-                }
-                Err(e) => return Completion::Throw(e),
-            }
-        }
-
-        // Module namespace: look up live binding from environment
-        {
-            let ns_data = self
-                .get_object_cell(obj_id)
-                .and_then(|obj| obj.borrow().module_namespace().cloned());
-            if let Some(ns_data) = ns_data {
-                // Deferred namespace: IsSymbolLikeNamespaceKey check
-                if ns_data.deferred
-                    && !Self::is_symbol_like_namespace_key(key, true)
-                    && let Err(e) = self.ensure_deferred_namespace_evaluation(obj_id)
-                {
-                    return Completion::Throw(e);
-                }
-                if let Some(binding_name) = ns_data.export_to_binding.get(key) {
-                    let module_path = ns_data.module_path.clone();
-                    match self.resolve_module_export_value(
-                        binding_name,
-                        &ns_data.env,
-                        module_path.as_deref(),
-                        key,
-                    ) {
-                        Ok(val) => return Completion::Normal(val),
-                        Err(e) => return Completion::Throw(e),
-                    }
-                }
-                // Fallback: check module's exports directly
-                if let Some(ref module_path) = ns_data.module_path
-                    && let Some(module) = self.module_registry_get(module_path)
-                    && let Some(val) = module.borrow().exports.get(key)
-                {
-                    return Completion::Normal(val.clone());
-                }
-            }
-        }
-
-        // Fallback: own property + prototype chain (for rare non-proxy, non-module cases)
-        let own_desc = if let Some(obj) = self.get_object_cell(obj_id) {
-            obj.borrow().get_own_property_full(key)
-        } else {
-            None
-        };
-        match own_desc {
-            Some(ref d) if d.get.is_some() && !matches!(d.get, Some(JsValue::Undefined)) => {
-                let getter = d.get.clone().unwrap();
-                self.call_function(&getter, this_val, &[])
-            }
-            Some(ref d) if d.get.is_some() => Completion::Normal(JsValue::Undefined),
-            Some(ref d) => Completion::Normal(d.value.clone().unwrap_or(JsValue::Undefined)),
-            None => {
-                let proto = if let Some(obj) = self.get_object_cell(obj_id) {
-                    obj.borrow().prototype_id
-                } else {
-                    None
-                };
-                if let Some(proto_rc) = proto {
-                    let proto_id = proto_rc;
-                    self.get_object_property(proto_id, key, this_val)
-                } else {
-                    Completion::Normal(JsValue::Undefined)
-                }
-            }
-        }
-    }
-
-    /// Proxy-aware [[HasProperty]] - checks proxy `has` trap, recurses on target if no trap.
-    pub(crate) fn proxy_has_property(&mut self, obj_id: u64, key: &str) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(key);
-            match self.invoke_proxy_trap(obj_id, "has", vec![target_val.clone(), key_val]) {
-                Ok(Some(v)) => {
-                    let trap_result = self.to_boolean_val(&v);
-                    if !trap_result
-                        && let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object_cell(t.id)
-                    {
-                        let target_desc = tobj.borrow().get_own_property(key);
-                        if let Some(ref desc) = target_desc {
-                            if desc.configurable == Some(false) {
-                                return Err(self.create_type_error(
-                                        "'has' on proxy: trap returned falsish for property which exists in the proxy target as non-configurable",
-                                    ));
-                            }
-                            if !tobj.borrow().extensible {
-                                return Err(self.create_type_error(
-                                        "'has' on proxy: trap returned falsish for property but the proxy target is not extensible",
-                                    ));
-                            }
-                        }
-                    }
-                    Ok(trap_result)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_has_property(t.id, key);
-                    }
-                    Ok(false)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object(obj_id) {
-            // Deferred namespace: trigger evaluation on [[HasProperty]] with non-symbol-like key
-            {
-                let is_deferred_ns = obj
-                    .borrow()
-                    .module_namespace()
-                    .as_ref()
-                    .is_some_and(|ns| ns.deferred);
-                if is_deferred_ns && !Self::is_symbol_like_namespace_key(key, true) {
-                    self.ensure_deferred_namespace_evaluation(obj_id)?;
-                }
-            }
-            // TypedArray §10.4.5.3 [[HasProperty]]: numeric indices handled by IsValidIntegerIndex only
-            {
-                let b = obj.borrow();
-                if b.typed_array_info().is_some()
-                    && let Some(index) =
-                        crate::interpreter::types::canonical_numeric_index_string(key)
-                {
-                    return Ok(is_valid_integer_index(b.typed_array_info().unwrap(), index));
-                }
-            }
-            if obj.borrow().has_own_property(key) {
-                return Ok(true);
-            }
-            // Walk prototype chain, checking for proxies
-            let proto = obj.borrow().prototype_id;
-            if let Some(proto_rc) = proto {
-                let proto_id = proto_rc;
-                return self.proxy_has_property(proto_id, key);
-            }
-            Ok(false)
-        } else {
-            Ok(false)
         }
     }
